Revert "[fix] changes location of statusviewcontroller"
